+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/Wishlist.orm.xml
@@ -40,7 +40,7 @@
             <join-column name="user_id" referenced-column-name="id" />
         </many-to-one>
 
-        <one-to-many field="items" target-entity="Sylius\Component\Core\Model\WishlistItemInterface" mapped-by="wishlist">
+        <one-to-many field="items" target-entity="Sylius\Component\Wishlist\Model\WishlistItemInterface" mapped-by="wishlist">
             <cascade>
                 <cascade-all />
             </cascade>
