+++ src/Sylius/Component/Wishlist/CHANGELOG.md
@@ -1,0 +1,6 @@
+CHANGELOG
+=========
+
+### v0.12.0
+
+* Initial dev release.
