+++ src/Sylius/Component/Wishlist/composer.json
@@ -1,0 +1,44 @@
+{
+    "name": "sylius/wishlist",
+    "type": "library",
+    "description": "Flexible wishlist component for PHP e-commerce projects.",
+    "keywords": ["wishlist", "watchlist", "shop", "webshop", "ecommerce", "sylius"],
+    "homepage": "http://sylius.org",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Paweł Jędrzejewski",
+            "homepage": "http://pjedrzejewski.com"
+        },
+        {
+            "name": "Sylius project",
+            "homepage": "http://sylius.org"
+        },
+        {
+            "name": "Community contributions",
+            "homepage": "http://github.com/Sylius/Sylius/contributors"
+        }
+    ],
+    "require": {
+        "php":             ">=5.3.3",
+
+        "sylius/resource": "0.12.*@dev"
+    },
+    "require-dev": {
+        "phpspec/phpspec": "~2.0"
+    },
+    "config": {
+        "bin-dir": "bin"
+    },
+    "autoload": {
+        "psr-4": { "Sylius\\Component\\Wishlist\\": "" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Sylius\\Component\\Wishlist\\spec\\": "spec/" }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "0.12-dev"
+        }
+    }
+}
